chore(parser): move getOperationErrors to common parser
